+++ changelog.md
@@ -1,3 +1,9 @@
 # Changelog
 
-Renamed 'folder-paths.js' to 'storage-paths.js'
+**./storage-paths.js**
+* Renamed function from 'defineFolderPaths'to 'defineStoragePaths'
+* Renamed existing properties:
+	* 'input' to 'inputFolder'
+	* 'output' to 'outputFolder'
+* Added 'optionsFile' property.
+	* Refers to 'options.json'